requirements.txt
customtkinter
Pillow
pandas
pyperclip
watchdog
pypng